Adds an option to check if the repository is aligned with the remote before the deploy
If in the .ini file the option'check_sync_with_remote' is set to true, it will be checked if the remote is aligned with the local repository. If it fails, the deploy will not be performed.
It can be useful when there is more than one person working on the repository, because it pervents inconsistent deploys to the production server.
If the remote is called 'origin', it is possible to specify the branch of the remote where to check the alignment, defining it through the option 'remote_branch'.